refactored the code that use model to do the recognition. Seperated the simulation mouse feature and gesture recognition feature. Added a dataset collector that can use to collect the gesture to train their model. (when you click the start capture, press "Space" to save the screenshot. Fix the python file naming format Integrated two features into app.py, now can use app.py file to start the application Add the new requirements.txt file for download the dependencies. Use the command “pip install requriments.txt" to do this.
