+++ app.py
@@ -1,14 +1,12 @@
 import platform
 import subprocess
 import tkinter as tk
-import cv2 as cv
-from setuptools import launch
-import utile as utile
-import camera
-import camera as camera
-from cleanup import cleanup
+from oldversion import utile as utile
+import gesture.mouse_simulator as ms
+import officialVersion.gesture_recognition as gs
+from oldversion import camera as ca
+from oldversion.cleanup import cleanup
 from tkinter import messagebox, ttk
-from PIL import Image, ImageTk
 
 # window = tk.Tk()
 # window.title("WAVEASE")
@@ -75,8 +73,13 @@ settings = {
 }
 
 
+def start_mouse_simulation():
+    ms.start_recognition()
+    messagebox.showinfo("message", "simulation closed!")
+
+
 def start_gesture_recognition():
-    camera.start_capture()
+    gs.start()
     messagebox.showinfo("message", "recognition closed!")
 
 
@@ -95,7 +98,7 @@ def open_settings():
     settings_window.geometry("300x250")
 
     # Camera and music app
-    avaible_camera = camera.find_available_cameras()
+    avaible_camera = ca.find_available_cameras()
     camera_options = avaible_camera
     music_app_options = ["App music 1", "Spotify 2"]
 
@@ -137,6 +140,7 @@ def exit_app():
     root.destroy()
     exit(0)
 
+
 def launch_app():
     spotify_path = utile.find_spotify_path()
     if spotify_path:
@@ -149,9 +153,6 @@ def launch_app():
         print("Spotify installation not found.")
 
 
-
-
-
 # Create the main windows
 root = tk.Tk()
 root.title("WavEase!")
@@ -160,15 +161,13 @@ root.minsize(500, 300)
 root.configure(background="#c3c3c3")
 
 # layout
-root.grid_columnconfigure([1,2], weight=1, minsize=70)
-#root.grid_columnconfigure(1, weight=1, minsize=70)
+root.grid_columnconfigure([1, 2], weight=1, minsize=70)
+# root.grid_columnconfigure(1, weight=1, minsize=70)
 
 root.grid_rowconfigure(0, weight=1, minsize=20)
 root.grid_rowconfigure(1, weight=1, minsize=20)
-#root.grid_rowconfigure([2,3], weight=1, minsize=10) ## does not work for Mac
-root.attributes('-alpha', 1.0) 
-
-
+# root.grid_rowconfigure([2,3], weight=1, minsize=10) ## does not work for Mac
+root.attributes('-alpha', 1.0)
 
 # Add a label
 label = tk.Label(root, text=""" __          __         _                                       _              __          __                  ______                        
@@ -184,8 +183,11 @@ label.grid(row=1, column=1, sticky="nsew", padx=8, pady=8, columnspan=2)
 start_button = tk.Button(root, text="Start Recognition", command=start_gesture_recognition)
 start_button.grid(row=2, column=1, padx=8, pady=8, ipadx=30, ipady=5)
 
+start_button = tk.Button(root, text="Start Mouse Simulation", command=start_mouse_simulation)
+start_button.grid(row=2, column=2, padx=8, pady=8, ipadx=30, ipady=5)
+
 launch_button = tk.Button(root, text="Launch app", command=launch_app)
-launch_button.grid(row=2, column=2, padx=8, pady=8, ipadx=30, ipady=5)
+launch_button.grid(row=2, column=3, padx=8, pady=8, ipadx=30, ipady=5)
 
 settings_button = tk.Button(root, text="Settings", command=open_settings)
 settings_button.grid(row=3, column=1, padx=8, pady=8, ipadx=30, ipady=5)
