+++ requirements.txt
@@ -1,16 +1,43 @@
-torch==1.13.0
-torchvision==0.14.0
-torchaudio==0.13.0
-pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
-timm==0.6.7
-ftfy
-regex
-einops
-fvcore
-eva-decord==0.6.1
-iopath
-numpy>=1.19
-matplotlib
-types-regex
-mayavi
-cartopy
+absl-py==2.1.0
+attrs==23.2.0
+autopy==4.0.0
+cffi==1.16.0
+comtypes==1.3.1
+contourpy==1.1.1
+cycler==0.12.1
+flatbuffers==24.3.25
+fonttools==4.50.0
+importlib_metadata==7.1.0
+importlib_resources==6.4.0
+jax==0.4.13
+kiwisolver==1.4.5
+matplotlib==3.7.5
+mediapipe==0.10.11
+ml-dtypes==0.2.0
+MouseInfo==0.1.3
+numpy==1.24.4
+opencv-contrib-python==4.9.0.80
+opencv-python==4.9.0.80
+opt-einsum==3.3.0
+packaging==24.0
+pillow==10.3.0
+protobuf==3.20.3
+psutil==5.9.8
+PyAutoGUI==0.9.54
+pycaw==20240210
+pycparser==2.22
+PyGetWindow==0.0.9
+PyMsgBox==1.0.9
+pynput==1.7.6
+pyparsing==3.1.2
+pyperclip==1.8.2
+PyRect==0.2.0
+PyScreeze==0.1.30
+python-dateutil==2.9.0.post0
+pytweening==1.2.0
+pywin32==306
+scipy==1.10.1
+screeninfo==0.8.1
+six==1.16.0
+sounddevice==0.4.6
+zipp==3.18.1
